MicrobolusViewModel
