+++ Loop/Views/Microboluses/MicrobolusView.swift
@@ -7,104 +7,11 @@
 //
 
 import SwiftUI
-import Combine
 import LoopCore
-import LoopKit
 import HealthKit
+import Combine
 
 struct MicrobolusView: View {
-    final class ViewModel: ObservableObject {
-        @Published fileprivate var microbolusesWithCOB: Bool
-        @Published fileprivate var microbolusesWithoutCOB: Bool
-        @Published fileprivate var partialApplication: Double
-        @Published fileprivate var microbolusesMinimumBolusSize: Double
-        @Published fileprivate var openBolusScreen: Bool
-        @Published fileprivate var disableByOverride: Bool
-        @Published fileprivate var lowerBound: String
-        @Published fileprivate var pickerMinimumBolusSizeIndex: Int
-        @Published fileprivate var partialApplicationIndex: Int
-        @Published fileprivate var event: String? = nil
-
-        // @ToDo: Should be able to get the to limit from the settings but for now defult to a low value
-        fileprivate let minimumBolusSizeValues = stride(from: 0.0, to: 0.51, by: 0.05).map { $0 }
-        fileprivate let partialApplicationValues = stride(from: 0.1, to: 1.01, by: 0.05).map { $0 }
-
-        private var cancellable: AnyCancellable!
-        fileprivate let formatter: NumberFormatter = {
-            let formatter = NumberFormatter()
-            formatter.numberStyle = .decimal
-            formatter.maximumFractionDigits = 1
-            return formatter
-        }()
-
-        fileprivate let unit: HKUnit
-
-        init(settings: Microbolus.Settings, glucoseUnit: HKUnit, eventPublisher: AnyPublisher<Microbolus.Event?, Never>? = nil) {
-            self.microbolusesWithCOB = settings.enabled
-            self.microbolusesWithoutCOB = settings.enabledWithoutCarbs
-            self.partialApplication = settings.partialApplication
-            self.microbolusesMinimumBolusSize = settings.minimumBolusSize
-            self.openBolusScreen = settings.shouldOpenBolusScreen
-            self.disableByOverride = settings.disableByOverride
-            self.lowerBound = formatter.string(from: settings.overrideLowerBound) ?? ""
-            self.unit = glucoseUnit
-
-            pickerMinimumBolusSizeIndex = minimumBolusSizeValues.firstIndex(of: Double(settings.minimumBolusSize)) ?? 0
-            partialApplicationIndex = partialApplicationValues.firstIndex(of: Double(settings.partialApplication)) ?? 0
-
-            let microbolusesMinimumBolusSizeCancellable = $pickerMinimumBolusSizeIndex
-                .map { Double(self.minimumBolusSizeValues[$0]) }
-                .sink { self.microbolusesMinimumBolusSize = $0 }
-
-            let partialApplicationCancellable = $partialApplicationIndex
-                .map { Double(self.partialApplicationValues[$0]) }
-                .sink { self.partialApplication = $0 }
-
-            let lastEventCancellable = eventPublisher?
-                .map { $0?.description }
-                .receive(on: DispatchQueue.main)
-                .sink { self.event = $0 }
-
-
-            cancellable = AnyCancellable {
-                microbolusesMinimumBolusSizeCancellable.cancel()
-                partialApplicationCancellable.cancel()
-                lastEventCancellable?.cancel()
-            }
-        }
-
-        func changes() -> AnyPublisher<Microbolus.Settings, Never> {
-            let lowerBoundPublisher = $lowerBound
-                .map { value -> Double in self.formatter.number(from: value)?.doubleValue ?? 0 }
-
-            return Publishers.CombineLatest(
-                Publishers.CombineLatest4(
-                    $microbolusesWithCOB,
-                    $microbolusesWithoutCOB,
-                    $partialApplication,
-                    $microbolusesMinimumBolusSize
-                ),
-                Publishers.CombineLatest3(
-                    $openBolusScreen,
-                    $disableByOverride,
-                    lowerBoundPublisher
-                )
-            )
-                .map {
-                    Microbolus.Settings(
-                        enabled: $0.0.0,
-                        enabledWithoutCarbs: $0.0.1,
-                        partialApplication: $0.0.2,
-                        minimumBolusSize: $0.0.3,
-                        shouldOpenBolusScreen: $0.1.0,
-                        disableByOverride: $0.1.1,
-                        overrideLowerBound: $0.1.2
-                    )
-                }
-                .eraseToAnyPublisher()
-        }
-    }
-
     @ObservedObject var viewModel: ViewModel
 
     init(viewModel: ViewModel) {
@@ -175,6 +82,7 @@ struct MicrobolusView: View {
                     TextField("0", text: $viewModel.lowerBound)
                     .keyboardType(.decimalPad)
                     .multilineTextAlignment(.trailing)
+                    .frame(height: 38)
 
                     Text(viewModel.unit.localizedShortUnitString)
                 }
@@ -212,7 +120,7 @@ struct MicrobolusView_Previews: PreviewProvider {
             )
         )
             .environment(\.colorScheme, .dark)
-            .previewLayout(.fixed(width: 375, height: 1300))
+            .previewLayout(.fixed(width: 375, height: 1000))
     }
 }
